Splitting profile out of admin
